Separate provider for openrouter
