+++ src/main/java/com/mulesoft/connectors/internal/models/ModelType.java
@@ -7,6 +7,7 @@ package com.mulesoft.connectors.internal.models;
 import com.mulesoft.connectors.internal.exception.error.ConfigValidationException;
 import com.mulesoft.connectors.internal.models.mistral.MistralModelName;
 import com.mulesoft.connectors.internal.models.openai.OpenAIModelName;
+import com.mulesoft.connectors.internal.models.openrouter.OpenRouterModelName;
 
 import java.util.Arrays;
 import java.util.stream.Stream;
