+++ src/main/java/com/mulesoft/connectors/internal/models/openrouter/providers/OpenRouterTextGenerationModelNameProvider.java
@@ -1,9 +1,9 @@
 /**
  * (c) 2003-2024 MuleSoft, Inc. The software in this package is published under the terms of the Commercial Free Software license V.1 a copy of which has been included with this distribution in the LICENSE.md file.
  */
-package com.mulesoft.connectors.internal.connection.openrouter;
+package com.mulesoft.connectors.internal.models.openrouter.providers;
 
-import com.mulesoft.connectors.internal.models.OpenRouterModelName;
+import com.mulesoft.connectors.internal.models.openrouter.OpenRouterModelName;
 import org.mule.runtime.api.value.Value;
 import org.mule.runtime.extension.api.values.ValueBuilder;
 import org.mule.runtime.extension.api.values.ValueProvider;
